8/13

+++ basic/Addition.java
@@ -2,9 +2,9 @@ package basic;
 
 public class Addition {
   public static void main(String args[]){
-    int a = Integer.parseInt(args[0]);
-    int b = Integer.parseInt(args[1]);
-    int sum = a + b;
+    double a = Double.parseDouble(args[0]);
+    double b = Double.parseDouble(args[1]);
+    double sum = a + b;
 
     System.out.println("答えは: " + sum);
   }
